cmsc491-vision/project-final/requirements.txt
configparser==3.5.0
cycler==0.10.0
decorator==4.3.0
kiwisolver==1.0.1
matplotlib==2.2.2
networkx==2.1
numpy==1.14.3
opencv-python==3.4.0.14
Pillow==5.1.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
